ItemTemplate 추가 아이템 사용 개선 치트 매니저 개선 TCHAR 컨버팅

+++ trunk/ScenePlayGame.cpp
@@ -126,7 +126,7 @@ void ScenePlayGame::update(float fDeltaTime)
         }
 
         //< 몬스터와 맵의 충돌 체크
-        for (size_t i = 0; i < MON_MGR->getSize(); i++)
+        for (int i = 0; i < (int)MON_MGR->getSize(); i++)
         {
             //< 주위의 선(벽) 찾기
             m_World->aroundLine(MON_MGR->getMonster(i)->getPos(), MON_MGR->getMonster(i)->getAroundVertex());
